app/requirements.txt
flask>=3.1.0
google-genai>=1.0.0
weasyprint>=60.0
python-dotenv>=1.0.0
markdown>=3.6
